+++ Loki/transposition.cpp
@@ -4,7 +4,7 @@
 TranspositionTable* tt = new TranspositionTable(TT_DEFAULT_SIZE);
 
 TranspositionTable::TranspositionTable(uint64_t size) {
-	numEntries = MB(size) / sizeof(TT_Entry);
+	numEntries = int(MB(size) / sizeof(TT_Entry));
 
 	entries = new TT_Entry[numEntries];
 
@@ -25,7 +25,7 @@ size_t TranspositionTable::size() {
 void TranspositionTable::resize(uint64_t size) {
 	delete[] entries;
 
-	numEntries = MB(size) / sizeof(TT_Entry);
+	numEntries = int(MB(size) / sizeof(TT_Entry));
 
 	entries = new TT_Entry[numEntries];
 
@@ -48,7 +48,7 @@ void TranspositionTable::clear_table() {
 
 // When probing the transposition table, we won't use time validating the entry while probing. Rather this will be done in search, so we'll just return the entry
 TT_Entry* TranspositionTable::probe_tt(uint64_t key, bool& hit) {
-	TT_Entry* slot = &entries[key & (numEntries - 1)];
+	TT_Entry* slot = &entries[key & (uint64_t(numEntries) - 1)];
 	uint64_t* data = (uint64_t*) &slot->data;
 
 	if (slot->key == (key ^ *data)) {
@@ -63,7 +63,7 @@ TT_Entry* TranspositionTable::probe_tt(uint64_t key, bool& hit) {
 
 // For now we are just using a replace all strategy
 void TranspositionTable::store_entry(const GameState_t* pos, int move, int score, int depth, int flag) {
-	TT_Entry* slot = &entries[pos->posKey & (numEntries - 1)];
+	TT_Entry* slot = &entries[pos->posKey & (uint64_t(numEntries) - 1)];
 	uint64_t* data = (uint64_t*) &slot->data;
 
 	assert(flag >= 0 && flag <= 2);
